requirements.txt
appdirs==1.4.4
asgiref==3.3.1
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
distlib==0.3.1
Django==3.1.3
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-js-asset==1.2.2
filelock==3.0.12
furl==2.1.0
httplib2==0.18.1
idna==2.10
importlib-metadata==2.0.0
Naked==0.1.31
orderedmultidict==1.0.1
psycopg2==2.8.6
python-http-client==3.3.1
pytz==2020.4
PyYAML==5.3.1
shellescape==3.8.1
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
zipp==3.4.0
gunicorn
whitenoise
dj-database-url
social-auth-app-django
python-decouple==3.3
Pillow==8.0.1
